Add support for rock and roll inputs
Rock and roll inputs are two-key sequences that result in a single behavior.
Rocks and rolls are distinguished from each other by the order in which the
keys are released. A rock occurs when the first key released is the second that
was pressed, and a roll occurs when the first key released is the first that
was pressed. These inputs are only captured for a specified group of keys and
override the keymap behavior for those keys. The rocknroll behavior table also
includes entries for single key presses, which are triggered on key release if
no other rocknroll key was pressed.